shadoker/requirements.txt
GitPython
pystache
orjson